make tray injectable
